requirements.txt
django==4.0.3
djangorestframework
django-oauth-toolkit
psycopg2-binary